Seperated Bestiary API into different projects and namespaces for domain and API(Some modules unfinished)
