+++ TakeInitiative.Bestiary.Domain/JSON/bestiary_Schema.json
