+++ TakeInitiative.sln
@@ -11,13 +11,16 @@ Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TakeInitiative.Api.Tests", 
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BestiaryAPI", "apps\TakeInitiative.Bestiary\BestiaryAPI.csproj", "{D3CFC8BC-E638-414E-A09E-802C06A8F803}"
 EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Bestiary", "Bestiary", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "Bestiary_DB_Builder", "apps\Bestiary_DB_Builder\Bestiary_DB_Builder.csproj", "{E261C2EF-C9CD-4D10-9499-774F60D8952B}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TakeInitiative.Bestiary.Domain", "TakeInitiative.Bestiary.Domain\TakeInitiative.Bestiary.Domain.csproj", "{365AEEB4-3B85-4FDA-BAA9-C3499D903397}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
 		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(ProjectConfigurationPlatforms) = postSolution
 		{745EE20A-57EE-4CE6-A090-01A2C373C9EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
@@ -32,10 +35,27 @@ Global
 		{D3CFC8BC-E638-414E-A09E-802C06A8F803}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{D3CFC8BC-E638-414E-A09E-802C06A8F803}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{D3CFC8BC-E638-414E-A09E-802C06A8F803}.Release|Any CPU.Build.0 = Release|Any CPU
+		{E261C2EF-C9CD-4D10-9499-774F60D8952B}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{E261C2EF-C9CD-4D10-9499-774F60D8952B}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{E261C2EF-C9CD-4D10-9499-774F60D8952B}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{E261C2EF-C9CD-4D10-9499-774F60D8952B}.Release|Any CPU.Build.0 = Release|Any CPU
+		{365AEEB4-3B85-4FDA-BAA9-C3499D903397}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{365AEEB4-3B85-4FDA-BAA9-C3499D903397}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{365AEEB4-3B85-4FDA-BAA9-C3499D903397}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{365AEEB4-3B85-4FDA-BAA9-C3499D903397}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
 	EndGlobalSection
 	GlobalSection(NestedProjects) = preSolution
 		{745EE20A-57EE-4CE6-A090-01A2C373C9EE} = {CFC275A3-AD62-4ABC-A90B-AFDA010134E3}
 		{EAE43750-0316-450C-B61B-3CCD023EF06E} = {CFC275A3-AD62-4ABC-A90B-AFDA010134E3}
-		{D3CFC8BC-E638-414E-A09E-802C06A8F803} = {CFC275A3-AD62-4ABC-A90B-AFDA010134E3}
+		{D3CFC8BC-E638-414E-A09E-802C06A8F803} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{02EA681E-C7D8-13C7-8484-4AC65E1B71E8} = {CFC275A3-AD62-4ABC-A90B-AFDA010134E3}
+		{E261C2EF-C9CD-4D10-9499-774F60D8952B} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+		{365AEEB4-3B85-4FDA-BAA9-C3499D903397} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {D1A836D7-FAF1-4473-AB33-47A5E9A6FEE4}
 	EndGlobalSection
 EndGlobal
