+++ apps/TakeInitiative.Bestiary/BestiaryAPI.csproj
@@ -12,4 +12,8 @@
     <PackageReference Include="Newtonsoft.Json" Version="13.0.3" />
   </ItemGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\..\TakeInitiative.Bestiary.Domain\TakeInitiative.Bestiary.Domain.csproj" />
+  </ItemGroup>
+
 </Project>
