+++ apps/TakeInitiative.Bestiary/BestiaryAPI.sln
@@ -1,0 +1,30 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio Version 17
+VisualStudioVersion = 17.14.36203.30
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "BestiaryAPI", "BestiaryAPI.csproj", "{A962241B-437F-1842-1C14-0FA4C80E3DF3}"
+EndProject
+Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "Solution Items", "Solution Items", "{8EC462FD-D22E-90A8-E5CE-7E832BA40C5D}"
+	ProjectSection(SolutionItems) = preProject
+		ImmuneResist.cs = ImmuneResist.cs
+	EndProjectSection
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{A962241B-437F-1842-1C14-0FA4C80E3DF3}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{A962241B-437F-1842-1C14-0FA4C80E3DF3}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{A962241B-437F-1842-1C14-0FA4C80E3DF3}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{A962241B-437F-1842-1C14-0FA4C80E3DF3}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {734B946A-08F6-4BDE-8722-E5D03B5698DC}
+	EndGlobalSection
+EndGlobal
